Upgrade to recommended Xcode settings and fix some build issues

+++ Apple2Mac/DDHidLib/lib/DDHidMouse.m
@@ -122,7 +122,7 @@
     return mButtonElements; 
 }
 
-- (unsigned) numberOfButtons;
+- (NSUInteger) numberOfButtons;
 {
     return [mButtonElements count];
 }
